requirements.txt
click==8.1.7
colorama==0.4.6
joblib==1.3.2
nltk==3.8.1
regex==2023.12.25
tqdm==4.66.2
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread